# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "电商风控规则引擎服务（grule），支持规则热更新",
        "title": "TCG-AI-ENGINE API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/risk/evaluate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "risk"
                ],
                "summary": "风控评估",
                "parameters": [
                    {
                        "description": "order/customer/product/merchant 四个 Fact",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tcg-ai-engine_internal_types_req.EvaluateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/tcg-ai-engine_internal_types_resp.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/tcg-ai-engine_internal_types_resp.EvaluateData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/rules": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "查看当前生效规则集",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/tcg-ai-engine_internal_types_resp.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/tcg-ai-engine_internal_engine.Info"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/rules/reload": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "手动触发规则热更新",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/tcg-ai-engine_internal_types_resp.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/tcg-ai-engine_internal_types_resp.ReloadData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "tcg-ai-engine_internal_engine.Info": {
            "type": "object",
            "properties": {
                "checksum": {
                    "description": "全量 GRL 内容的 SHA-256，热更新判据",
                    "type": "string"
                },
                "loaded_at": {
                    "description": "本版本生效时间",
                    "type": "string"
                },
                "rule_count": {
                    "description": "规则条数（rule 级别，非文件数）",
                    "type": "integer"
                },
                "rule_names": {
                    "description": "所有规则名，字典序",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "source": {
                    "description": "规则来源：file:rules / oracle:RISK_RULES",
                    "type": "string"
                }
            }
        },
        "tcg-ai-engine_internal_model.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                }
            }
        },
        "tcg-ai-engine_internal_model.Customer": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/tcg-ai-engine_internal_model.Address"
                },
                "blacklisted": {
                    "description": "是否在黑名单",
                    "type": "boolean"
                },
                "chargeback_count": {
                    "description": "拒付（chargeback）次数",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "description": "会员等级 1~5",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "points": {
                    "description": "当前积分",
                    "type": "integer"
                },
                "refund_count": {
                    "description": "历史退款次数",
                    "type": "integer"
                },
                "register_days": {
                    "description": "注册天数",
                    "type": "integer"
                },
                "tags": {
                    "description": "客户标签，GRL 里通过 Customer.HasTag(\"xxx\") 查询",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total_orders": {
                    "description": "历史订单数",
                    "type": "integer"
                }
            }
        },
        "tcg-ai-engine_internal_model.Merchant": {
            "type": "object",
            "properties": {
                "blacklisted": {
                    "description": "是否在黑名单",
                    "type": "boolean"
                },
                "category": {
                    "description": "经营类目",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "description": "评分 0~5",
                    "type": "number"
                },
                "refund_rate": {
                    "description": "退款率 0~1",
                    "type": "number"
                },
                "years_active": {
                    "description": "经营年数",
                    "type": "integer"
                }
            }
        },
        "tcg-ai-engine_internal_model.Order": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "订单金额",
                    "type": "number"
                },
                "channel": {
                    "description": "下单渠道：app / web / h5",
                    "type": "string"
                },
                "device_id": {
                    "description": "设备指纹，空值视为风险信号",
                    "type": "string"
                },
                "freight": {
                    "description": "基础运费报价（规则可减免/加收，结果写入 Result.Freight）",
                    "type": "number"
                },
                "hour_of_day": {
                    "description": "下单时刻（0~23，简化处理避免 time.Time）",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "ip_country": {
                    "description": "下单 IP 归属国，空表示未知",
                    "type": "string"
                },
                "is_first_order": {
                    "description": "是否首单",
                    "type": "boolean"
                },
                "pay_method": {
                    "description": "支付方式：credit_card / cod / balance ...",
                    "type": "string"
                },
                "quantity": {
                    "description": "购买件数",
                    "type": "integer"
                }
            }
        },
        "tcg-ai-engine_internal_model.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "类目：electronics / gift_card / luxury / books ...",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_virtual": {
                    "description": "虚拟商品（卡密、点卡等），无物流、易套现",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "presale_days": {
                    "description": "预售天数，0 表示现货",
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "restricted": {
                    "description": "管制/限购商品",
                    "type": "boolean"
                },
                "stock": {
                    "description": "库存",
                    "type": "integer"
                }
            }
        },
        "tcg-ai-engine_internal_model.Result": {
            "type": "object",
            "properties": {
                "decision": {
                    "description": "approve / review / reject",
                    "type": "string"
                },
                "discount": {
                    "description": "营销（互斥组：一单只命中一个折扣档位）",
                    "type": "number"
                },
                "discount_name": {
                    "description": "命中的折扣档位，空串=未定，作为互斥闸门",
                    "type": "string"
                },
                "extra_discount": {
                    "description": "叠加折扣是否已生效（防重复叠加）",
                    "type": "boolean"
                },
                "final_amount": {
                    "description": "结算与追踪",
                    "type": "number"
                },
                "freight": {
                    "description": "运费",
                    "type": "number"
                },
                "freight_free": {
                    "description": "免运费闸门：免运费后不再加收偏远地区附加费",
                    "type": "boolean"
                },
                "hit_rules": {
                    "description": "命中的规则名，按执行顺序",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "points_earned": {
                    "description": "Finalize 时按倍率结算",
                    "type": "integer"
                },
                "points_rate": {
                    "description": "积分（互斥组：积分倍率只定一次）",
                    "type": "number"
                },
                "reject_reason": {
                    "description": "拒单原因",
                    "type": "string"
                },
                "risk_score": {
                    "description": "风控",
                    "type": "number"
                }
            }
        },
        "tcg-ai-engine_internal_types_req.EvaluateRequest": {
            "type": "object",
            "properties": {
                "customer": {
                    "$ref": "#/definitions/tcg-ai-engine_internal_model.Customer"
                },
                "merchant": {
                    "$ref": "#/definitions/tcg-ai-engine_internal_model.Merchant"
                },
                "order": {
                    "$ref": "#/definitions/tcg-ai-engine_internal_model.Order"
                },
                "product": {
                    "$ref": "#/definitions/tcg-ai-engine_internal_model.Product"
                }
            }
        },
        "tcg-ai-engine_internal_types_resp.EngineMeta": {
            "type": "object",
            "properties": {
                "checksum": {
                    "type": "string"
                },
                "rule_count": {
                    "type": "integer"
                }
            }
        },
        "tcg-ai-engine_internal_types_resp.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "0 成功，非 0 失败",
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "tcg-ai-engine_internal_types_resp.EvaluateData": {
            "type": "object",
            "properties": {
                "engine": {
                    "$ref": "#/definitions/tcg-ai-engine_internal_types_resp.EngineMeta"
                },
                "order_id": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/tcg-ai-engine_internal_model.Result"
                }
            }
        },
        "tcg-ai-engine_internal_types_resp.ReloadData": {
            "type": "object",
            "properties": {
                "changed": {
                    "description": "false = 内容没变，未重建",
                    "type": "boolean"
                },
                "info": {
                    "$ref": "#/definitions/tcg-ai-engine_internal_engine.Info"
                }
            }
        }
    }
}